requirements.txt
streamlit 
geopandas 
pandas 
osmnx 
folium 
shapely 
python-pptx